Factor out basic http server class
